+++ SimuladorCredito/DTO/Responses/Parcela.cs
@@ -3,7 +3,7 @@
 public class Parcela
 {
     public int numero { get; set; }
-    public double valorAmortizacao { get; set; }
-    public double valorJuros { get; set; }
-    public double valorPrestacao { get; set; }
+    public decimal valorAmortizacao { get; set; }
+    public decimal valorJuros { get; set; }
+    public decimal valorPrestacao { get; set; }
 }
