+++ SimuladorCredito/DTO/Responses/RespostaSimutacao.cs
@@ -5,6 +5,6 @@ public class RespostaSimutacao
     public int idSimulacao { get; set; }
     public int codigoProduto { get; set; }
     public string descricaoProduto { get; set; }
-    public double taxaJuros { get; set; }
+    public decimal taxaJuros { get; set; }
     public List<ResultadoSimulacao> resultadoSimulacao { get; set; }
 }
